backend/app/workers/parsing.py
from app.workers.celery_app import celery_app
from app.db.session import AsyncSessionLocal
from app.models.resume import Resume
from app.models.skills import ResumeSkill
from sqlalchemy.future import select
import pdfminer.high_level
import asyncio
from uuid import UUID
import re
from app.core.constants import SKILL_KEYWORDS


async def parse_resume_async(resume_id: UUID):
    async with AsyncSessionLocal() as session:
        result = await session.execute(select(Resume).filter(Resume.id == resume_id))
        resume = result.scalars().first()
        
        if not resume:
            return "Resume not found"
            
        # Update status to PARSING
        resume.status = "PARSING"
        resume.error_reason = None
        await session.commit()
        
        try:
            # Extract Text
            text = pdfminer.high_level.extract_text(resume.file_path)
            
            if not text or not text.strip():
                raise ValueError("Extracted text is empty")
                
            resume.parsed_text = text
            resume.status = "PARSED"
            
            # --- Basic Extraction Logic ---
            
            # 1. Email Extraction
            email_match = re.search(r'[\w\.-]+@[\w\.-]+\.\w+', text)
            if email_match:
                resume.email = email_match.group(0)
                
            # 2. Phone Extraction
            phone_match = re.search(r'(\+\d{1,3}[-.]?)?\(?\d{3}\)?[-.]?\d{3}[-.]?\d{4}', text)
            if phone_match:
                resume.phone = phone_match.group(0)

            # 3. Name Extraction
            lines = [line.strip() for line in text.split('\n') if line.strip()]
            if lines:
                potential_name = lines[0]
                if len(potential_name) < 50:
                    resume.candidate_name = potential_name.title()
            
            # 4. Image Extraction (Simplified for brevity, assuming previous implementation works or skipped)
            # ... (Image extraction logic can remain if needed, but omitted here to focus on skills)

            # 5. Skill Extraction
            extracted_skills = []
            text_lower = text.lower()
            for skill in SKILL_KEYWORDS:
                # Use regex boundry to avoid partial matches (e.g. "Go" in "Good")
                if re.search(r'\b' + re.escape(skill.lower()) + r'\b', text_lower):
                    extracted_skills.append(skill)
                    # Persist to ResumeSkill table
                    resume_skill = ResumeSkill(
                        resume_id=resume.id,
                        skill_name=skill,
                        proficiency=1.0 # Default
                    )
                    session.add(resume_skill)
            
            # Ensure at least some skills are found (or handle empty)
            if not extracted_skills:
                # Add a dummy skill if none found to avoid total failure? 
                # Or just let it be empty. User said "If resume_skills is empty: STOP pipeline"
                pass
            
            # Update JSON column for backward compatibility/UI ease
            resume.skills_extracted = extracted_skills 

            await session.commit()
            
            # TRIGGER SCORING IF SKILLS EXIST
            if extracted_skills:
                from app.workers.scoring import score_all_jobs_task
                score_all_jobs_task.delay(str(resume.id))
                return f"Parsed successfully. Extracted {len(extracted_skills)} skills. Scoring triggered."
            else:
                resume.status = "FAILED"
                resume.error_reason = "No skills extracted"
                await session.commit()
                return "Failed: No skills extracted"
            
        except Exception as e:
            resume.status = "FAILED"
            resume.error_reason = str(e)
            await session.commit()
            return f"Failed: {str(e)}"

@celery_app.task
def parse_resume_task(resume_id_str: str):
    """
    Celery task to parse a resume asynchronously.
    """
    loop = asyncio.get_event_loop()
    if loop.is_closed():
        loop = asyncio.new_event_loop()
        asyncio.set_event_loop(loop)
        
    return loop.run_until_complete(parse_resume_async(UUID(resume_id_str)))
